Lots of tweaking and tuning. Extensions and improvements to the README. Fixed some bugs in 'surveyor_remode' involving the transformation matrix hand-off. Code clean-up to remove debug statements and tidy up functionality; lots of work still to do on that front.
Still have an issue with de-noising the final point cloud; looks to be very similar to the output of REMODE before the de-noising step referred to at the end of the UZH-RPG video showcasing the full system. Need to reach out to the UZH people to see if they can assist...as far as I can tell, there is no documentation on what that final 'de-noising' step entails.